Fixed target framework and mono posix version
