+++ src/EventStore.ClientAPI.Embedded/EventStore.ClientAPI.Embedded.csproj
@@ -1,26 +1,39 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net47</TargetFramework>
+    <TargetFramework>net471</TargetFramework>
     <Platform>x64</Platform>
     <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
     <DebugType>portable</DebugType>
     <DebugSymbols>true</DebugSymbols>
-    <OutputPath>..\..\bin\$(Configuration)\$(MSBuildProjectName)\</OutputPath>
-    <IntermediateOutputPath>..\..\obj\$(Configuration)\$(MSBuildProjectName)\</IntermediateOutputPath>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
     <DocumentationFile>$(OutputPath)\EventStore.ClientAPI.Embedded.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\EventStore.ClientAPI\EventStore.ClientAPI.csproj" />
     <ProjectReference Include="..\EventStore.Common\EventStore.Common.csproj" />
     <ProjectReference Include="..\EventStore.Core\EventStore.Core.csproj" />
-    <ProjectReference Include="..\EventStore.Projections.Core\EventStore.Projections.Core.csproj" />  
+    <ProjectReference Include="..\EventStore.Projections.Core\EventStore.Projections.Core.csproj" />
   </ItemGroup>
   <!-- TODO(jen20): Investigate whcih of these are correct -->
-  <Target Name="CopyDependencies" BeforeTargets="BeforeBuild">
-    <Copy ContinueOnError="ErrorAndStop" SourceFiles="..\libs\x64\linux\libjs1.so" DestinationFolder="$(OutDir)" Condition="'$(OS)' != 'Windows_NT' And '$(IsMac)' == 'false'" />
-    <Copy ContinueOnError="ErrorAndStop" SourceFiles="..\libs\x64\mac\libjs1.dylib" DestinationFolder="$(OutDir)" Condition="'$(OS)' != 'Windows_NT' And '$(IsMac)' == 'true'" />
-    <Copy ContinueOnError="ErrorAndStop" SourceFiles="..\libs\x64\win\js1.dll" DestinationFolder="$(OutDir)" Condition=" '$(OS)' == 'Windows_NT' " />
-    <Copy ContinueOnError="ErrorAndStop" SourceFiles="..\libs\x64\win\js1.pdb" DestinationFolder="$(OutDir)" Condition=" '$(OS)' == 'Windows_NT' " />
+  <Target Name="CopyDependencies"
+          BeforeTargets="BeforeBuild">
+    <Copy ContinueOnError="ErrorAndStop"
+          SourceFiles="..\libs\x64\linux\libjs1.so"
+          DestinationFolder="$(OutDir)"
+          Condition="'$(OS)' != 'Windows_NT' And '$(IsMac)' == 'false'" />
+    <Copy ContinueOnError="ErrorAndStop"
+          SourceFiles="..\libs\x64\mac\libjs1.dylib"
+          DestinationFolder="$(OutDir)"
+          Condition="'$(OS)' != 'Windows_NT' And '$(IsMac)' == 'true'" />
+    <Copy ContinueOnError="ErrorAndStop"
+          SourceFiles="..\libs\x64\win\js1.dll"
+          DestinationFolder="$(OutDir)"
+          Condition=" '$(OS)' == 'Windows_NT' " />
+    <Copy ContinueOnError="ErrorAndStop"
+          SourceFiles="..\libs\x64\win\js1.pdb"
+          DestinationFolder="$(OutDir)" />
+              
   </Target>
 </Project>
