+++ src/EventStore.Common/EventStore.Common.csproj
@@ -1,18 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>net471</TargetFramework>
     <Platform>x64</Platform>
     <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
     <DebugType>portable</DebugType>
     <DebugSymbols>true</DebugSymbols>
-    <OutputPath>..\..\bin\$(Configuration)\$(MSBuildProjectName)\</OutputPath>
-    <IntermediateOutputPath>..\..\obj\$(Configuration)\$(MSBuildProjectName)\</IntermediateOutputPath>
   </PropertyGroup>
   <ItemGroup>
       <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
       <PackageReference Include="NLog" Version="4.5.10" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\EventStore.Common.Utils\EventStore.Common.Utils.csproj" />
     <ProjectReference Include="..\EventStore.Rags\EventStore.Rags.csproj" />
   </ItemGroup>
 </Project>
