+++ src/EventStore.Native/TransactionLog/Unbuffered/NativeFileWindows.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.ComponentModel;
 using System.IO;
-using EventStore.Common.Utils;
 using Microsoft.Win32.SafeHandles;
 
 namespace EventStore.Core.TransactionLog.Unbuffered
