+++ src/EventStore.sln
@@ -35,7 +35,14 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EventStore.Rags", "EventSto
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EventStore.ClientAPI.Embedded", "EventStore.ClientAPI.Embedded\EventStore.ClientAPI.Embedded.csproj", "{C1239448-07E3-4E26-A8B2-AF9687E99E7C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventStore.Native", "EventStore.Native\EventStore.Native.csproj", "{03FA0763-6D41-44C0-911D-662AE2B10084}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EventStore.Native", "EventStore.Native\EventStore.Native.csproj", "{03FA0763-6D41-44C0-911D-662AE2B10084}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A19631BB-65CF-4726-8732-4EFD977E9B32}"
+	ProjectSection(SolutionItems) = preProject
+		Directory.Build.props = Directory.Build.props
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EventStore.Common.Utils", "EventStore.Common.Utils\EventStore.Common.Utils.csproj", "{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -251,6 +258,18 @@ Global
 		{03FA0763-6D41-44C0-911D-662AE2B10084}.Release|x64.Build.0 = Release|Any CPU
 		{03FA0763-6D41-44C0-911D-662AE2B10084}.Release|x86.ActiveCfg = Release|Any CPU
 		{03FA0763-6D41-44C0-911D-662AE2B10084}.Release|x86.Build.0 = Release|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Debug|x64.Build.0 = Debug|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Debug|x86.Build.0 = Debug|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Release|Any CPU.Build.0 = Release|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Release|x64.ActiveCfg = Release|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Release|x64.Build.0 = Release|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Release|x86.ActiveCfg = Release|Any CPU
+		{647867C2-54FA-4C2C-9BCD-7F85A4A8F787}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
